+++ API/src/users/Controller.ts
@@ -61,11 +61,20 @@ usersController.get("/valide/:bool", async (req, res) => {
 
 usersController.delete("/:id", async (req, res) => {
     try{
-        await UserRepository.deleteUser(req.params.id);
-        res.status(200).send({
-            status: 200,
-            message: "Deleted"
-        })
+        const user = await UserRepository.getUserById(req.params.id);
+
+        if (user) {
+            await UserRepository.deleteUser(req.params.id);
+            res.status(200).send({
+                status: 200,
+                message: "Deleted"
+            })
+        }else{
+            res.status(404).send({
+                status: 404,
+                message: "Not found!"
+            })
+        }
     }catch(error){
         res.status(404).send({
             status: 404,
@@ -77,12 +86,14 @@ usersController.delete("/:id", async (req, res) => {
 usersController.post("/", async (req, res) => {
 
     try{
+        debug(req.body);
         const user = await UserRepository.createUser(req.body);
         res.status(200).send(user);
     }catch(error){
         res.status(400).send({
             status: 400,
-            message: "Bad user data"
+            message: "Bad user data",
+            error: error
         })
     }
